refactor: 메인화면 상세화면 CSS 수정

+++ src/pages/posting/Posting.js
@@ -81,12 +81,14 @@ const Posting = () => {
   };
 
 	const onRemoveQuestionHandler = () => {
-		dispatch(asyncRemoveQuestion({
-			questionId: question.id,
-			userId: userData.id
-		}));
-		alert('글이 삭제되었습니다.');
-		navigate('/', { state: { userData: userData }});
+		// alert('글이 삭제되었습니다.');
+		if (window.confirm('삭제하시겠습니까?')) {
+			dispatch(asyncRemoveQuestion({
+				questionId: question.id,
+				userId: userData.id
+			}));
+			navigate('/', { state: { userData: userData }});
+		}
 	}
